DirtyWeaver/requirements.txt
Pillow
numpy
scipy
argon2-cffi
cryptography
tqdm
pysoundfile